[New sql update for realmd 3859_spell_affect.sql] 1. Implemented 64bit SpellFamilyFlags 2. Fixed tooltips of spells that are affected by SpellMods
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@3859 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/ObjectMgr.cpp
@@ -2701,7 +2701,7 @@ void ObjectMgr::LoadSpellAffects()
         sa.Category = fields[4].GetUInt16();
         sa.SkillId = fields[5].GetUInt16();
         sa.SpellFamily = fields[6].GetUInt8();
-        sa.SpellFamilyMask = fields[7].GetUInt32();
+        sa.SpellFamilyMask = fields[7].GetUInt64();
         sa.Charges = fields[8].GetUInt16();
 
         SpellAffect.insert(SpellAffectMap::value_type(entry,sa));
